String Tokenizer
This will accept a string, convert it into a vector of ints and perform an operation on it. Please note returning vector by value is optimised way to do it as this will involve move semantics. Returning a vector does not copy the object, thus avoiding wasting extra speed/space. Under the hood, it points the pointer to the returned vector object.